feat: create foundation for syntax highlighting
This is a bit of a mess; I'm going to try to start by getting it working
and then refactor like mad to get it into a state I actually *like*.

Current state: once I actually load a syntax set (including, maybe, a
built-in one? Syntect's docs need some love...), this should support the
no-highlighting and highlight-everything scenarios, but not yet the much
more complicated just-tag-it scenario which is what I *actually* want.
Still, that's actually enough to start refactoring a bit. Maybe put a
'TBD' message in there and call it good and move on to actually starting
to get closer to a point where I can build up some of the *other* pieces
of this story.