Throw away almost everything and start over
I realized something incredibly important: there is no reason whatsoever
to try to make this do everything. There are lots of other generators
out there which do that, and not one of them serves *my* needs well. So:
throw out all my original thoughts on making this hyper-configurable,
and just build exactly the tool I need. Perhaps it'll be useful to
others and perhaps not, but that isn't the point!