Start on footnote emit
- Collect the backrefs while emitting footnote references into the obdy
  (the actually emit there is not yet implemented, but the types and
  design are).
- Given a collection of backrefs, emit either a single one with `↩` or
  multiple with `↩<sup>{index}</sup>`.
- Add utilities for emitting links, since that will be useful in many
  places.

Note that these are also not the final form of the HTML (it needs some
further a11y work and also classes to be applied), but do represent a
good start on it.

---

This basically just uses standalone functions. I'm sort of thinking on
my preferred style: `impl`s on items have the nice feature (the *very*
nice feature!) of being able to do `.call()` and that also gives nice
IDE integration, but it's not always clear that that is actually better
in every case; in particular, doing it *consistently* in a case like
this one actually *requires* introducing wrapper types throughout, so
that `to_html()` calls can be appropriately parameterized, and then
avoiding overhead with `#[repr(transparent)]` or some such.