Implement `Iterator` and `IntoInterator` for `second_pass::State`
I actually only need `IntoIterator` for this, and so may end up just
deleting the `Iterator` implementation later -- but I wanted to think
through how to do this. Earlier, I thought I might be able to do it in
terms of `std::vec::Iter`'s `Iterator` implementation, just using the
`chain` combinator, but the lifetimes go a bit wild and I could not see
a way to do it.