Naively implement 'Rendered' metadata
Just run a dumb Markdown implementation on them. This seems fine? Because
I do not care to do a bunch of fancy footnote processing on those things.
However, I *should* make the option configuration happen in a single spot
at some point, presumably the `markdown` module in a lazy_cell or such.