Reimplement Markdown metadata and rewrite handling
The previous approach committed a grievous sin: it coupled the Markdown
parsing itself to the details of metadata handling and rewriting. This
also ended up scattering around all sorts of noise and nonsense when it
came to where and how errors were handled. Instead of entirely wrapping
up all the Markdown handling in a single pass, make the two-phase work
explicit.

Keep the `first_pass` and `second_pass` details private, by providing
metadata as a `String` to the caller who drives the passes, but only an
opaque `struct` for the content required by the second pass, so that the
internal details are never exposed. This is a nicely modular pattern,
and it decreases the total coupling required for the parsing and
rendering *substantially* as well as eliminating coupling to the
`Metadata` module entirely.

With this done, and a better handle on what my error-handling structure
should be in the first place, substantially improve that too.