A working first pass at JSON Schema for items
This would need to be tweaked considerably for actually using these for
validation, but this is a good starting point, and I learned enough
about JSON Schema to be able to use it for this going forward if I so
desire.

Still the big open question: do I want to drive this whole thing via the
piles-of-files on disk model, or via a database? If the latter, a lot of
this goes away in favor of database schemas (which have their own
issues!) but referencing things just becomes... referencing things.

One thing this does suggest/support more easily is the idea of things
like a `library.lx.yaml` file where I just update *data* (and have it
validated both dynamically in-editor and programmatically via the build
process) and can reference its contents in other contexts. (How that
would work, exactly, is an open question!)