Introduce third party Atom feed definition
I am not sure whether I will keep this or just implement Atom myself,
with some variety of templating, the same way I did with my 11ty-based
site. The downside to the latter: I messed it up a *bunch* then! The
usual upside: I have end-to-end control. The other consideration: if I
do use this third-party library, I need to make sure that the items I
provide to it have been further "munged" to insert the various headings
etc. that I use (Assumed Audience etc.). If I build it myself with a
template language, I get the ability to just do all of that "for free".