Actually get text-only rewrites working
This is not my favorite approach ever, but it does work. It is, as
expected, slightly slower than doing a single post-processing pass on
items *after* finishing, likely in part because of cloning the Tera
instance. However, it accomplishes what I need it to: it does not touch
the code blocks!