Split out Markdown into its own crate
This is the first part of #1. It introduces the new crate and updates
all the references to it, and stubs out the binary that will call it
later. This is *not* what I want the final form here to be, _per se_,
since callers need to *always* pass in syntax sets and a `rewrite`,
even when that does not make sense to them, but my initial pass at
making those optional (or otherwise handling it) went... poorly? And
I therefore chose to set that aside for tonight in favor of making
progress that I could *land*.