Implement lx-md end to end
Current status: it doesn't quite work the way I want it to yet, because
my `render` pipeline assumes that there will always be metadata! The
next step here will be to figure out how I want to drive that. I think
the basic move will be to pull that part of the state machine up a
level, because it is not really a *Markdown* error for it not to be
there anyway: it is a *Page* error.

(I may also want to change how the syntax rewriting is handled: just
making it optional was fine, but it definitely does add a higher degree
of coupling by making the Markdown renderer actually Syntect-aware.)

Happily, though, this does actually work, which is neat!

squash w/parent pls