Update Markdown handling to not *require* metadata
The basic two-pass render works the same way, but metadata is optional,
so that files where I do not have it can be rendered just the same as
in the case where I do. Missing metadata is a problem for `lx`, but it
is *not* a problem for (e.g.) `lx-md`.