Fold `lx-md` into `lx`
This will make it much simpler to manage all of this going forward! When
I call it from e.g. BBEdit, I can just do `lx md` instead of `lx-md`,
and the same from the command line. This is not especially
well-structured, and I expect I will do some work to refactor it later.
However, it does the job of solving the specific issue I set out to
solve here, and I am going to make a habit of taking exactly that
approach to make continuous forward progress on this.