Configure toolchain w/profile and components
Use the `minimal` profile, but adding in rustfmt and clippy. (This is
close to what `default` gives, but does not install `rust-docs`.)

This does not put the file in the root of the repo, so it may not work
as desired in terms of the GitHub Actions config; the only good way to
find out is to try it. 🤪