Use Syntect built-in SyntaxSet
This takes the time for rendering the 'jj init' post from 136ms down
to 20ms. This is *not* the final approach I want to take, because I
ultimately want to bake in my own syntaxes. Probably, anyway: I do not
think Syntect ships with *everything* I want, but I do not actually
know that for sure; I will have to check! In any case, this handily
solves the perf issue for the moment.