Finish removing `ugh` by splitting it into its constituents
This does not *quite* finish moving to a component-ized model; for that
I will also need to finish moving all the `.span-*` stuff etc. over to
mixins and defining classes from them. It does, however, get me to a
point where all of that initial hodgepodge is in a coherent place, and
thus to a good foundation for the remaining componentization work!