Get the music page layout basically working
This has a bit of stuff to clean up around how the row spanning is done
for the YouTube embeds at the top, but it is actually working at this
point, at least at a basic level. I will take it! With some minor tweaks
to footer and to the header layout, I can launch this, possibly as early
as tomorrow!