lx: get the live reloading part working 🎉
This does not yet do rebuilds, and the client does not skip reloading
based on whether the paths matter (and in fact it may not even make
sense to bother with that), but: good progress!

Supporting changes:

- Enable Axum’s WebSocket feature.
- Add `serde_json`.
- Update to Rust 1.79, because I want to use the fancy async function in
  trait feature for my own local `race` definition.
- Add (for now!) a test-only `index.html` that lets me demo the behavior
  I want for live reloading.

As a bonus, swap out `serde_derive` for just enabling serde’s `derive`
feature, and update all the imports accordingly.