lx: Swap out Tera for minijinja
Both of these are *not 100% what I want*… but minijinja is *more* what I
want *for now*, in that it is actually compatible with Jinja2 (so the
docs for Jinja2 are useful for it) and all my existing knowledge comes
over cleanly. While yes, in the fullness of time, I would *love* to
write some kind of templating engine of my own, the reality is that this
is probably going to be it for the foreseeable future. Might as well be
something I have higher confidence in and comfort with.