lx: better error handling in rewrite callback
`Page::build` and the whole chain it kicks off can be thread safe as
long as the errors propagated along the way are themselves thread safe.
This means a lot of `+ Send + Sync` annotations along the way! However,
it makes for much cleaner error handling.

In a follow on change, I would like to convert all of this over from
`Box<dyn std::error::Error + Send + Sync>` to [`anyhow::Error`][a]: it
is cheaper (one word instead of two) and bakes in the constraints for
thread safety, including the `'static` that I am *implicitly* depending
on in the current implementation.

[a]: https://docs.rs/anyhow/latest/anyhow/struct.Error.html

(Notably, this was possibly *before* the switch from Tera to minijinja,
but I am better at Rust than I was a year ago and did not realize how to
wrangle the plumbing then. Today… it was pretty easy.)