music: expand site with minimal link/info pages for works
This is not how I want this to be built long-term, of course, and it is
not even how I want these pages to look/work long-term. It will do the
job for right now, though, and that is what counts: it lets me get away
from the ticking bomb that is Songwhip.