v6: for now, do not `@use` the highlighting CSS
I need to think through how to approach these! Something akin to the
approach I took for v5 is probably broad-strokes the right thing to do,
i.e. using a mixin or some such, but in any case I *know* that I do not
want it to be just slapped directly in the way it was in the previous
commit.