lx: start building out template rendering capability
This hardcodes a bunch of things, of course, but/and I can tweak all of
that later. The point right now is: *motion*. This successfully renders
the “jj init” post at a very basic level: no styles, nothing but an HTML
`<title>` and a `<body>` which has the content rendered directly into
it, but that is more than I had a day ago!