lx: Render basic output with cascade and layouts
This correctly resolves the cascade, including requiring `lx.data.yaml`
as the location of any data file in a directory. It also resolves a
number of items in the cascade more correctly, e.g. merging fields where
it makes sense to do so (lists of tags, partial values for qualifiers,
and so on). Finally, with that in place it is able to use the `layout`
chosen by either an item or the nearest part of the data cascade to
render the page in question.

This is actually a sufficient foundation that I could already render all
of `chriskrycho.music`/`music.chriskrycho.com` with some hard-coding for
the CSS build! That is suggestive, though: the next move is probably to
get `grass` wired up so that it is not to hard-code in.