lx: support for musical works in metadata
Implement full support in the `serial` and `metadata` layers, including
merging the two correctly. Along the way, rework the implementation of
`Cascade::find_map` to deal in references so I *can* return references
in the future if that is useful. (Right now everything just clones the
values out of the cascade directly, and that is reasonable and likely to
remain the case, but it is nice to have that ability.)