lx: iterate on design for data vs. other pages
In the interest of allowing arbitrary pages to be defined simply with a
`<page>.lx.yaml` file, a directory’s data must now be at `_data.yaml` if
it exists. When getting site files, distinguish content files from data
files using that distinction. For now, still only *build* Markdown files
(since that is all I have support for), but the wiring should be easy to
add for e.g. `feeds/email.yaml` or similar. This should solve most or
all of my need for custom pages without needing to do additional parsing
on them, given the ability to name a specific template. (If it turns out
I *do* need additional capabilities, I can add it then, of course.)