lx: *very* basic (partly broken) JSON Feed item support
There is actually enough data being correctly munged end to end on the
`Page` item that this is *mostly* possible now. The `TODO`s on this are
good signals of what I still need to wire up. This is also a good “tell”
for the fact that I need to think on exactly how I am going to render
JSON feeds! `serde_json` is probably a good bet but it is definitely not
obvious how that integrates with the existing system I have outlined for
myself with templates and page files. Something to figure out Soon™.