lx: support for shared infrastructure, e.g. `head`
Make it so any site can load from both `_shared` and `sites/<the site>`
for `_ui` and `_styles`. This makes it possible to extract things like a
common handler for the `head` (though this will almost certainly want
more data, e.g. the site config!) to be able to have a single place I do
that for *all* of these sites.