lx: fix page title and URL rendering
To make this work, I need to change how a `Page` defines its root URL.
The previous approach had it dynamically constructed on demand using the
input URL. That never made a lot of sense, and I broadly knew that; but
now it is time to actually fix it! The move here is to incorporate that
info into the `Page` itself, so that it is always available anywhere the
`Page` is available.

Accordingly, eagerly compute the relative path so URLs can be checked
*without* needing to pass the input directly around later. To that end,
separate out the end of the markdown rendering from the final
construction of the `Page` type, so the chain is now `Prepared ->
Rendered -> Page`, where the final step includes *all* of the data
required to represent a `Page` thereafter. (I think I can simplify this
even further in a follow-on “tidy” step: a lot of this data does not
need to be carried around *at all*.)

To make that work, additionally update `RootedPath` to be the source of
truth for the logic around paths: previously that was attached to the
`Page` type, as part of its lazy logic around URL handling, but that was
the wrong home for it and I confused myself with it a few times along
the way.

With all of those supporting changes in place, make a working `url_for`
helper function and invoke it in the `head` template.