tidy(lx): do not pass references to `source` through `page` chain
This way only the types and functions that actually *require* that info
ever get a reference to it. The main distinction here is that `Prepared`
and `Rendered` do not need to know or care about it directly, although
the `prepare` function gets it as an argument because `Metadata` needs
it for building the `slug`. This puts a slightly higher burden on the
`build` function to keep the references around… but it also means some
of the `PathBuf` clones that were happening every time to keep ownership
clean now *only* happen in error cases! All in all, a nice little tidy.